+++ EatPic-iOS/Sources/Screens/Home/Calender/CalenderCard/CalenderCardView.swift
@@ -27,6 +27,7 @@ struct CalenderCardView: View {
                 Spacer().frame(height: 43)
             }
             
+            
             // 레시피 링크 ~ 메모 등의 버튼 뷰 4개
             ZStack {
                 // VStack은 기본적으로
