+++ EatPic-iOS/Sources/Screens/Home/HomeView.swift
@@ -15,7 +15,7 @@ struct HomeView: View {
                 
                 MealStatusView()
                 
-                RecomPicCardView()
+                RecomPicCardHomeView()
                 
                 MyBadgeStatusHomeView()
