+++ EatPic-iOS/Sources/Screens/Home/ViewModels/MealStatusViewModel.swift
@@ -10,8 +10,8 @@ import Foundation
 class MealStatusViewModel: ObservableObject {
     
     @Published var mealStatus: [MealStatusModel] = [
-        MealStatusModel(mealTime: "아침", isRecorded: true, imageName: "Home/img1"),
-        MealStatusModel(mealTime: "점심", isRecorded: true, imageName: "Home/img1"),
+        MealStatusModel(mealTime: "아침", isRecorded: true, imageName: "Home/exampleSalad"),
+        MealStatusModel(mealTime: "점심", isRecorded: true, imageName: "Home/exampleSalad"),
         MealStatusModel(mealTime: "저녁", isRecorded: false), // imageName 없음
         MealStatusModel(mealTime: "간식", isRecorded: false)  
     ]
